added tool bar
