+++ src/Navigation/index.js
@@ -3,7 +3,6 @@ import {NavigationContainer} from '@react-navigation/native';
 import {createStackNavigator} from '@react-navigation/stack';
 import {View, Text} from 'react-native';
 import HomeScreen from '../screens/Home';
-
 const Stack = createStackNavigator();
 
 export const LearnScreen = () => (
@@ -14,8 +13,14 @@ export const LearnScreen = () => (
 const HomeStack = () => {
   return (
     <NavigationContainer>
-      <Stack.Navigator>
-        <Stack.Screen name="Home" component={HomeScreen} />
+      <Stack.Navigator headerMode="false">
+        <Stack.Screen
+          name="Home"
+          component={HomeScreen}
+          // options={{
+          //   headerLeft: props => <Header />,
+          // }}
+        />
         <Stack.Screen name="Learn" component={LearnScreen} />
       </Stack.Navigator>
     </NavigationContainer>
