+++ storybook/storyLoader.js
@@ -12,6 +12,7 @@ function loadStories() {
   require('../src/components/NextButton/NextButton.stories');
   require('../src/components/SectionHeading/SectionHeading.stories');
   require('../src/components/Textarea/Textarea.stories');
+  require('../src/components/ToolBar/ToolBar.stories');
   require('../src/components/ToolButton/ToolButton.stories');
   require('./stories/Button/Button.stories');
   require('./stories/Welcome/Welcome.stories');
@@ -26,6 +27,7 @@ const stories = [
   '../src/components/NextButton/NextButton.stories',
   '../src/components/SectionHeading/SectionHeading.stories',
   '../src/components/Textarea/Textarea.stories',
+  '../src/components/ToolBar/ToolBar.stories',
   '../src/components/ToolButton/ToolButton.stories',
   './stories/Button/Button.stories',
   './stories/Welcome/Welcome.stories',
